linting and restructure
