+++ app/utils/sparql-utils.ts
@@ -1,0 +1,56 @@
+import { optionMapOr } from 'mow-registry/utils/option';
+export const sparqlEscapeString = (value: string) =>
+  '"""' + value.replace(/[\\"]/g, (match) => '\\' + match) + '"""';
+
+export const sparqlEscapeUri = (value: string) => {
+  return (
+    '<' +
+    value.replace(/[\\"<>]/g, function (match) {
+      return '\\' + match;
+    }) +
+    '>'
+  );
+};
+
+export function sparqlEscapeBool(value) {
+  return value ? '"true"^^xsd:boolean' : '"false"^^xsd:boolean';
+}
+
+export function sparqlEscapeDateTime(value) {
+  return '"' + new Date(value).toISOString() + '"^^xsd:dateTime';
+}
+
+export async function executeQuery<Binding = Record<string, RDF.Term>>({
+  query,
+  endpoint,
+  abortSignal,
+}: QueryConfig) {
+  const encodedQuery = encodeURIComponent(query.trim());
+
+  const response = await fetch(endpoint, {
+    method: 'POST',
+    mode: 'cors',
+    headers: {
+      Accept: 'application/sparql-results+json',
+      'Content-Type': 'application/x-www-form-urlencoded; charset=UTF-8',
+    },
+    body: `query=${encodedQuery}`,
+    signal: abortSignal,
+  });
+
+  if (response.ok) {
+    return response.json() as Promise<QueryResult<Binding>>;
+  } else {
+    throw new Error(
+      `Request to ${endpoint} was unsuccessful: [${response.status}] ${response.statusText}`,
+    );
+  }
+}
+
+export async function executeCountQuery(queryConfig: QueryConfig) {
+  const response = await executeQuery<{ count: { value: string } }>(
+    queryConfig,
+  );
+
+  return optionMapOr(0, parseInt, response.results.bindings[0]?.count.value);
+}
